# Optimize marshalling