--- a/tests/e2e/v1/mod.rs
+++ b/tests/e2e/v1/mod.rs
@@ -1,0 +1,1 @@
+mod func;