--- a/wasmi_v1/src/engine/mod.rs
+++ b/wasmi_v1/src/engine/mod.rs
@@ -346,7 +346,6 @@ impl EngineInner {
         }
     }
 
-    #[inline(always)]
     fn execute_host_func<C>(
         &mut self,
         mut ctx: C,