--- a/wasmi_v1/src/engine/value_stack.rs
+++ b/wasmi_v1/src/engine/value_stack.rs
@@ -301,6 +301,10 @@ impl ValueStack {
         self.entries[self.stack_ptr]
     }
 
+    pub fn drop(&mut self, depth: usize) {
+        self.stack_ptr -= depth;
+    }
+
     /// Pops the last [`StackEntry`] from the [`ValueStack`] as `T`.
     pub fn pop_as<T>(&mut self) -> T
     where
@@ -414,19 +418,11 @@ impl ValueStack {
         &self.entries[0..len]
     }
 
-    /// Pops the last `depth` stack entries and returns them as slice.
-    ///
-    /// Stack entries are returned in the order in which they got pushed
-    /// onto the value stack.
-    ///
-    /// # Panics
-    ///
-    /// If the value stack does not have at least `depth` stack entries.
-    pub fn pop_as_slice(&mut self, depth: usize) -> &[StackEntry] {
-        self.stack_ptr -= depth;
-        let start = self.stack_ptr;
-        let end = self.stack_ptr + depth;
-        &self.entries[start..end]
+    /// Returns an exclusive slice to the last `depth` entries in the value stack.
+    pub fn peek_as_slice_mut(&mut self, depth: usize) -> &mut [StackEntry] {
+        let start = self.stack_ptr - depth;
+        let end = self.stack_ptr;
+        &mut self.entries[start..end]
     }
 
     /// Clears the [`ValueStack`] entirely.