--- a/wasmi_v1/src/foreach_tuple.rs
+++ b/wasmi_v1/src/foreach_tuple.rs
@@ -1,4 +1,3 @@
-
 /// Macro to help implement generic trait implementations for tuple types.
 macro_rules! for_each_tuple {
     ($mac:ident) => {