--- a/wasmi_v1/src/func/into_func.rs
+++ b/wasmi_v1/src/func/into_func.rs
@@ -214,13 +214,13 @@ macro_rules! for_each_tuple {
         $mac!( 7 T1 T2 T3 T4 T5 T6 T7);
         $mac!( 8 T1 T2 T3 T4 T5 T6 T7 T8);
         $mac!( 9 T1 T2 T3 T4 T5 T6 T7 T8 T9);
-        // $mac!(10 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10);
-        // $mac!(11 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11);
-        // $mac!(12 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12);
-        // $mac!(13 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12 T13);
-        // $mac!(14 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12 T13 T14);
-        // $mac!(15 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12 T13 T14 T15);
-        // $mac!(16 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12 T13 T14 T15 T16);
+        $mac!(10 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10);
+        $mac!(11 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11);
+        $mac!(12 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12);
+        $mac!(13 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12 T13);
+        $mac!(14 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12 T13 T14);
+        $mac!(15 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12 T13 T14 T15);
+        $mac!(16 T1 T2 T3 T4 T5 T6 T7 T8 T9 T10 T11 T12 T13 T14 T15 T16);
     }
 }