--- a/wasmi_v1/src/func/into_func.rs
+++ b/wasmi_v1/src/func/into_func.rs
@@ -4,6 +4,7 @@ use super::{
 };
 use crate::{
     engine::{ReadParams, WriteResults},
+    foreach_tuple::for_each_tuple,
     Caller,
     SignatureEntity,
 };