This PR optimizes marshalling of function parameters and return values when calling host functions from Wasm execution.
This basically tries to go the optimal path where the host function parameters are read out directly from the value stack and the host function directly writes results back into the value stack.
The benchmarks show improvements between 10-15%:

# Note

- Host calls when executed from the host side now use the `v1` interpreter engine, too.
- The optimized marshalling mainly got improved for the common case where Wasm calls into the host.
- This PR implementation allows for an efficient `TypedFunc` implementation. (https://github.com/paritytech/wasmi/issues/303)
- There no longer is the need for the artificial `scratch` buffer for the `v1` engine type so this PR removes it again.

# ToDo

- [x] Use more declarative macros instead of hand coding some generic trait implementations.
- [x] Write some unit tests to guard some edge cases for host calls.
    - [ ] Fix failing tests.

# Benchmarks

## Before

```
execute/host_calls/v0   time:   [75.558 us 75.773 us 76.006 us]                                  
                        change: [-1.0263% -0.2417% +0.5489%] (p = 0.54 > 0.05)
execute/host_calls/v1   time:   [59.512 us 60.099 us 61.050 us]                                  
                        change: [-4.4018% -3.6043% -2.7357%] (p = 0.00 < 0.05)
```

## After

```
execute/host_calls/v0   time:   [73.723 us 75.298 us 77.553 us]                                  
                        change: [-2.7991% -1.3959% +0.2272%] (p = 0.06 > 0.05)
execute/host_calls/v1   time:   [52.763 us 52.959 us 53.169 us]                                  
                        change: [-2.8487% -1.7205% -0.7480%] (p = 0.00 < 0.05)
```